Extract permalink into component
