+++ src/components/character.tsx
@@ -1,6 +1,5 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { useEffect, useState } from "react";
-import Button from "react-bootstrap/Button";
 import { FaDAndD, FaDiceD20 } from "react-icons/fa";
 import {
   GiBloodySword,
@@ -62,10 +61,9 @@ import { getLanguages } from "../utils/getLanguages";
 import { SavedCharacterData } from "../utils/getSavedCharacterData";
 import { getSpells } from "../utils/getSpells";
 import { getTraits } from "../utils/getTraits";
-import { saveCharacterData } from "../utils/saveCharacterData";
 import parchment from "../static/parchment.png";
-import { useHistory } from "react-router-dom";
 import StyledItemsForPurchase from "./itemsForPurchase";
+import Permalink from "./characterSheetComponents/permalink";
 
 interface Props {
   abilityScores: number[];
@@ -84,7 +82,6 @@ const CharacterImpl: React.SFC<ImplProps> = ({
   includeKnaveSpells,
   savedCharacterData
 }) => {
-    let history = useHistory();
   // Hit Points
   const [hitPoints, setHitPoints] = useState(
     (savedCharacterData && savedCharacterData.hitPoints) ||
@@ -201,52 +198,16 @@ const CharacterImpl: React.SFC<ImplProps> = ({
 
   return (
     <div className={className}>
-      {!savedCharacterData ? (
-        <PermalinkButtonContainer>
-          <PermalinkButton
-            style={{ width: "275px" }}
-            variant="outline-secondary"
-            onClick={() => {
-              let URL = saveCharacterData(characterName,
-                                classSelection,
-                                traits,
-                                abilityScores,
-                                hitPoints,
-                                languages,
-                                spells,
-                                equipment.characterEquipmentString,
-                                equipment.slotsToFill,
-                                includeKnaveSpells
-                              );
-              history.push(`/savedCharacter/1&${URL}`);
-            }}
-          >
-            Permalink
-          </PermalinkButton>
-          <div>(or hit back to see other class options for current ability scores)</div>
-        </PermalinkButtonContainer>
-      ) : (
-        <SaveMessageContainer>
-          <SaveHeader>
-            <FontAwesomeIcon
-              icon={"exclamation-triangle"}
-              size="sm"
-              style={{ margin: "0.5rem" }}
-            />
-            IMPORTANT
-            <FontAwesomeIcon
-              icon={"exclamation-triangle"}
-              size="sm"
-              style={{ margin: "0.5rem" }}
-            />
-          </SaveHeader>
-          <SaveMessage
-            dangerouslySetInnerHTML={createMarkup(
-              "<strong>Bookmark this page. Save the results!\n</strong>"
-            )}
-          />
-        </SaveMessageContainer>
-      )}
+      <Permalink savedCharacterData={savedCharacterData} 
+                 characterName={characterName} 
+                 classSelection={classSelection} 
+                 traits={traits} 
+                 abilityScores={abilityScores} 
+                 hitPoints={hitPoints} 
+                 languages={languages} 
+                 spells={spells} 
+                 equipment={equipment} 
+                 includeKnaveSpells={includeKnaveSpells}/>
       <CharacterName>{characterName}</CharacterName>
       <ClassTitle>
         {`Level 1 ${characterClasses[classSelection].name}`}
@@ -968,67 +929,6 @@ const CharacterImpl: React.SFC<ImplProps> = ({
     </div>
   );
 };
-
-const SaveMessage = styled.div`
-  display: flex;
-  justify-content: center;
-  border-bottom: 1px solid black;
-  white-space: pre-line;
-`;
-
-const SaveHeader = styled.div`
-  display: flex;
-  justify-content: center;
-  font-family: "Sancreek", cursive;
-  font-size: 2rem;
-`;
-
-const SaveMessageContainer = styled.div`
-  @keyframes fadeOut {
-    0% {
-      opacity: 1;
-      height: 73px;
-    }
-    25% {
-      opacity: 0.75;
-      height: 73x;
-    }
-    50% {
-      opacity: 0.5;
-      height: 73px;
-    }
-    75% {
-      opacity: 0.25;
-      height: 73px;
-    }
-    95% {
-      opacity: 0;
-      height: 73px;
-    }
-    100% {
-      opacity: 0;
-      height: 0px;
-    }
-  }
-  -webkit-animation: fadeOut 8s;
-  animation: fadeOut 8s;
-  -webkit-animation-fill-mode: forwards;
-  animation-fill-mode: forwards;
-  -webkit-animation-timing-function: ease-in-out;
-  animation-timing-function: ease-in-out;
-`;
-
-const PermalinkButton = styled(Button)`
-  color: black;
-`;
-
-const PermalinkButtonContainer = styled.div`
-  justify-content: center;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  padding: 1rem 0 0.5rem 0;
-`;
 
 const CharacterName = styled.div`
   display: flex;
